+++ R/reexports.R
@@ -1,3 +1,23 @@
 #' @importFrom basejump calculateMetrics
 #' @export
 basejump::calculateMetrics
+
+#' @importFrom cowplot draw_plot
+#' @export
+cowplot::draw_plot
+
+#' @importFrom cowplot ggdraw
+#' @export
+cowplot::ggdraw
+
+#' @importFrom scales log_breaks
+#' @export
+scales::log_breaks
+
+#' @importFrom cowplot plot_grid
+#' @export
+cowplot::plot_grid
+
+#' @importFrom scales pretty_breaks
+#' @export
+scales::pretty_breaks